+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the convention described at
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Changed
+- Index format changed significantly, indices must be rebuilt.
+
 ## [0.1.0] - 2019-08-27
 ### Changed
 - Initial release